removed pinMode for pin 4 added line breaks for a new file and the file end discoverd the meaning of static and removed the keyword 'new' changed some logical data types fixed GPSPeriod calculation added boolean to replace colon with dot because of file naming

+++ lib/GPX/GPX.h
@@ -9,7 +9,7 @@ class GPX {
     void reuseLastGpxOrCreateNew(const char* currentDateTime);
     void createNewGpxFile(const char* currentDateTime);
     void writeEndOfFile();
-    const char* createNewTrackPoint(double latitude, double longitude, double altitude, const char* time, const char* fix, int hdop, double speed);
+    const char* createNewTrackPoint(double latitude, double longitude, double altitude, const char* time, const char* fix, int hdop, int speed);
     void appendCurrentGpxFile(const char* str);
 };
 #endif  // __GPX_H__
